move content box component
