+++ src/components/CreateDialog/SummaryStep.tsx
@@ -14,7 +14,7 @@ import { useAuth } from "components/AuthContext";
 import Loader from "components/Loader";
 import { StepProps } from "./types";
 import InfoBox from "components/InfoBox";
-import ContentBox from "./ContentBox";
+import ContentBox from "../ContentBox";
 
 function SummaryStep({
   active,
